executor/requirements.txt
celery[redis]==4.*
pandas
numpy
six
omegaconf
pyyaml
orjson>=3,<4
psycopg2
